Changed animation to interactive mode, moved initialisation into SimulationVariables
